Pull Standard Implementation of increaseAge into the StandardAgingItem sub class since the aging rules are different for every type of aging item
